Memory optimization
Saving around 322 bytes of flash and 59 bytes of RAM.
However, this will limit the histogram on a 256 chars area on the screen.

+++ LcdHistogram.h
@@ -2,7 +2,7 @@
  * LcdHistogram
  *
  * LcdHistogram.h
- * Pierre Faivre - oct. 2014
+ * Pierre Faivre - oct. 2014 - dec. 2016
  */
 
 #ifndef LcdHistogram_h
@@ -36,7 +36,7 @@ enum GraphMode {
 
 class LcdHistogram {
     public:
-        LcdHistogram(LiquidCrystal* lcd, int posx, int posy, int length);
+        LcdHistogram(LiquidCrystal* lcd, byte posx, byte posy, byte length);
         ~LcdHistogram();
         
         void push(float number);
@@ -47,9 +47,9 @@ class LcdHistogram {
         
     private:
         LiquidCrystal* _lcd; // LCD to print the graph
-        int _posx; // Row on the screen of the first bar
-        int _posy; // Line on the screen of the first bar
-        int _length; // Length of the histogram in characters
+        byte _posx; // Row on the screen of the first bar
+        byte _posy; // Line on the screen of the first bar
+        byte _length; // Length of the histogram in characters
         float* _data; // Numbers to plot
 
         GraphMode _mode; // Current mode of drawing (default : DYNAMIC)
